feeds some more bugs to my refrigerator (#6358)

+++ code/controllers/subsystem/lighting.dm
@@ -38,7 +38,7 @@ SUBSYSTEM_DEF(lighting)
 		i += 1
 
 		var/datum/light_source/L = queue[i]
-		L.update_corners()
+		L?.update_corners()
 		if(!QDELETED(L))
 			L.needs_update = LIGHTING_NO_UPDATE
 		else
@@ -66,8 +66,9 @@ SUBSYSTEM_DEF(lighting)
 		i += 1
 
 		var/datum/lighting_corner/C = queue[i]
-		C.needs_update = FALSE //update_objects() can call qdel if the corner is storing no data
-		C.update_objects()
+		if(!QDELETED(C))
+			C.needs_update = FALSE //update_objects() can call qdel if the corner is storing no data
+			C.update_objects()
 
 		// We unroll TICK_CHECK here so we can clear out the queue to ensure any removals/additions when sleeping don't fuck us
 		if(init_tick_checks)
